refactor: update landing page components and remove unused files
- Replaced the CenterCard and ClientInteractivity components with new implementations for improved functionality and design.
- Updated the Home component to utilize AnimationProvider, BackgroundLines, and LeaderCard for enhanced visual effects.
- Removed deprecated components such as FloatingEarlyAccessChat, FloatingEarlyAccessClient, and related early access forms to streamline the codebase.
- Adjusted the early access form structure and styling for better user engagement and clarity.
- Enhanced CSS styles for landing page elements to improve overall aesthetics and responsiveness.
