+++ apps/www/src/components/landing/application-view/3d-modeling-card.tsx
@@ -2,17 +2,17 @@
 
 import { useEffect, useRef, useState } from "react";
 
-import type { IntegrationCategory } from "../../../lib/landing/constants";
+import type { IntegrationCategory } from "../../../config/landing";
 
-interface VideoVFXCardProps {
+interface ThreeDModelingCardProps {
   category: IntegrationCategory;
 }
 
-export const VideoVFXCard = ({ category }: VideoVFXCardProps) => {
+export const ThreeDModelingCard = ({ category }: ThreeDModelingCardProps) => {
   const [isHovered, setIsHovered] = useState(false);
   const cardRef = useRef<HTMLDivElement>(null);
 
-  // Mouse tracking for lens flare effect
+  // Mouse tracking for 3D effects
   useEffect(() => {
     const element = cardRef.current;
     if (!element || !isHovered) return;
@@ -22,8 +22,8 @@ export const VideoVFXCard = ({ category }: VideoVFXCardProps) => {
       const x = (e.clientX - rect.left) / rect.width;
       const y = (e.clientY - rect.top) / rect.height;
 
-      element.style.setProperty("--mouse-x", `${x * 100}%`);
-      element.style.setProperty("--mouse-y", `${y * 100}%`);
+      element.style.setProperty("--mouse-x", `${(x - 0.5) * 20}deg`);
+      element.style.setProperty("--mouse-y", `${(0.5 - y) * 20}deg`);
     };
 
     element.addEventListener("mousemove", handleMouseMove);
@@ -33,67 +33,67 @@ export const VideoVFXCard = ({ category }: VideoVFXCardProps) => {
   return (
     <div
       ref={cardRef}
-      className={`relative h-full w-full overflow-hidden border p-6 transition-all duration-400 ease-out ${
+      className={`relative overflow-hidden border p-6 transition-all duration-400 ease-out ${
         isHovered
-          ? "border-yellow-400 shadow-[0_0_30px_rgba(255,215,0,0.2)]"
+          ? "border-cyan-400 shadow-[0_0_30px_rgba(0,255,255,0.3)]"
           : "border-white/10 bg-white/[0.02]"
       }`}
       onMouseEnter={() => setIsHovered(true)}
       onMouseLeave={() => setIsHovered(false)}
       style={
         {
-          "--mouse-x": "50%",
-          "--mouse-y": "50%",
+          "--mouse-x": "0deg",
+          "--mouse-y": "0deg",
           ...(isHovered && {
-            background: "linear-gradient(135deg, #1a1a1a 0%, #2a1810 100%)",
+            background: "linear-gradient(135deg, #0f1419 0%, #1a2332 100%)",
+            transform: `rotateX(var(--mouse-y, 0deg)) rotateY(var(--mouse-x, 0deg))`,
+            transformStyle: "preserve-3d",
           }),
         } as React.CSSProperties
       }
     >
-      {/* Film Effects Background */}
+      {/* 3D Background Elements */}
       {isHovered && (
         <div
-          className="absolute inset-0 opacity-0"
+          className="animate-in fade-in absolute inset-0 opacity-0 duration-400"
           style={{ animation: "fadeIn 0.4s ease-out forwards" }}
         >
-          {/* Film Grain */}
           <div
-            className="absolute inset-0 opacity-10 mix-blend-overlay"
+            className="absolute top-1/2 right-5 h-15 w-15"
             style={{
-              backgroundImage: `url("data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTAwIiBoZWlnaHQ9IjEwMCIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIj48ZGVmcz48ZmlsdGVyIGlkPSJub2lzZSI+PGZlVHVyYnVsZW5jZSBiYXNlRnJlcXVlbmN5PSIwLjkiIG51bU9jdGF2ZXM9IjQiIHNlZWQ9IjIiLz48L2ZpbHRlcj48L2RlZnM+PHJlY3Qgd2lkdGg9IjEwMCUiIGhlaWdodD0iMTAwJSIgZmlsdGVyPSJ1cmwoI25vaXNlKSIgb3BhY2l0eT0iMC4xIi8+PC9zdmc+")`,
-            }}
-          />
-          {/* Film Strip */}
-          <div
-            className="absolute top-0 bottom-0 left-0 opacity-30"
-            style={{
-              width: "20px",
-              background:
-                "linear-gradient(180deg, #ffd700 0%, transparent 50%, #ffd700 100%)",
+              width: "60px",
+              height: "60px",
+              transformStyle: "preserve-3d",
+              animation: "rotate3d 20s infinite linear",
             }}
           >
-            {["10%", "30%", "50%", "70%"].map((top, index) => (
+            {[
+              { name: "front", transform: "rotateY(0deg) translateZ(30px)" },
+              { name: "back", transform: "rotateY(180deg) translateZ(30px)" },
+              { name: "right", transform: "rotateY(90deg) translateZ(30px)" },
+              { name: "left", transform: "rotateY(-90deg) translateZ(30px)" },
+              { name: "top", transform: "rotateX(90deg) translateZ(30px)" },
+              { name: "bottom", transform: "rotateX(-90deg) translateZ(30px)" },
+            ].map((face) => (
               <div
-                key={index}
-                className="absolute h-2 w-2 rounded-full bg-black"
+                key={face.name}
+                className="absolute h-15 w-15 border border-cyan-400/60 bg-cyan-400/10"
                 style={{
-                  top,
-                  left: "6px",
+                  width: "60px",
+                  height: "60px",
+                  transform: face.transform,
                 }}
               />
             ))}
           </div>
-          {/* Lens Flare */}
           <div
-            className="pointer-events-none absolute h-25 w-25 blur-sm"
+            className="absolute inset-0 opacity-30"
             style={{
-              width: "100px",
-              height: "100px",
-              left: "var(--mouse-x, 50%)",
-              top: "var(--mouse-y, 50%)",
-              transform: "translate(-50%, -50%)",
-              background:
-                "radial-gradient(circle, rgba(255, 215, 0, 0.3) 0%, transparent 70%)",
+              backgroundImage: `
+                linear-gradient(rgba(0, 255, 255, 0.1) 1px, transparent 1px),
+                linear-gradient(90deg, rgba(0, 255, 255, 0.1) 1px, transparent 1px)
+              `,
+              backgroundSize: "20px 20px",
             }}
           />
         </div>
@@ -112,14 +112,14 @@ export const VideoVFXCard = ({ category }: VideoVFXCardProps) => {
                 key={app.name}
                 className={`flex items-center gap-2 rounded-md border px-3 py-2 transition-all duration-300 ${
                   isHovered
-                    ? "border-yellow-400/30 bg-yellow-400/10 shadow-[0_0_10px_rgba(255,215,0,0.2)]"
+                    ? "border-cyan-400/30 bg-cyan-400/10"
                     : "border-transparent bg-white/5"
                 }`}
                 style={
                   {
                     "--delay": `${index * 0.1}s`,
                     ...(isHovered && {
-                      transform: `translateY(${-3 * index * 0.1}px)`,
+                      transform: `translateZ(10px) translateY(${-5 * index * 0.1}px)`,
                       transitionDelay: `${index * 0.1}s`,
                     }),
                   } as React.CSSProperties
@@ -132,8 +132,7 @@ export const VideoVFXCard = ({ category }: VideoVFXCardProps) => {
                     className="h-5 w-5 object-contain opacity-80 transition-all duration-300"
                     style={{
                       ...(isHovered && {
-                        filter:
-                          "sepia(1) hue-rotate(30deg) saturate(2) brightness(1.2)",
+                        filter: "drop-shadow(0 0 5px rgba(0, 255, 255, 0.5))",
                       }),
                     }}
                   />
