--- apps/www/src/components/landing/page-header.tsx
@@ -1,10 +1,0 @@
-export const PageHeader = () => {
-  return (
-    <div className="optimized-page-header">
-      <h2 className="text-foreground mb-4 text-3xl font-bold">
-        Works with your
-        <span className="text-primary ml-2 italic">favorite tools</span>
-      </h2>
-    </div>
-  );
-};
