--- apps/www/src/lib/animation/index.ts
@@ -1,16 +1,0 @@
-/**
- * Animation utilities and physics calculations
- * Re-export all animation-related utilities from a single entry point
- */
-
-// Animation constants and timing (reads from CSS)
-export * from "./constants";
-
-// CSS timing utilities
-export * from "./css-timing";
-
-// Validation utilities
-export * from "./utils";
-
-// Pure spring physics calculations
-export * from "./spring-physics";
